backend-flash-dockpoly/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
starlette==0.38.6

# Database
sqlalchemy==2.0.36
alembic==1.17.2
asyncpg==0.31.0
psycopg[binary]==3.3.2
pymysql==1.1.0

# Pydantic & Validation
pydantic==2.12.4
pydantic-settings==2.12.0
email-validator==2.2.0

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
cryptography==43.0.3

# Admin Panel
fastadmin[fastapi,sqlalchemy]==0.2.22

# File Handling
python-multipart==0.0.20
aiofiles==24.1.0

# PDF Generation (simpler alternatives)
fpdf==1.7.2
pypdf==5.2.0
reportlab==4.2.5

# Document Generation
python-docx==1.1.2
xlsxwriter==3.2.0

# HTTP Client
httpx==0.27.2

# Environment
python-dotenv==1.0.1
